# Conflicts:
#	protected/modules/OphCiExamination/assets/css/module.css
#	protected/modules/OphCiExamination/widgets/js/HistoryMedications.js
#	protected/modules/OphCiExamination/widgets/views/HistoryMedicationsEntry_event_edit.php

Fix migration to import only drugs into the 'Formulary' set

+++ protected/modules/OphCoCorrespondence/controllers/AdminController.php
@@ -221,28 +221,25 @@ class AdminController extends \ModuleAdminController
             Audit::add('admin', 'view', $macro->id, null, array('module' => 'OphCoCorrespondence', 'model' => 'LetterMacro'));
         }
 
-        $letter_type_gp_id = \Yii::app()->db->createCommand()->select('id')->from('ophcocorrespondence_letter_recipient')->where('name=:name', array(':name' => 'GP'))->queryScalar();
-        $letter_type_patient_id = \Yii::app()->db->createCommand()->select('id')->from('ophcocorrespondence_letter_recipient')->where('name=:name', array(':name' => 'Patient'))->queryScalar();
         $this->render('_macro', array(
             'macro' => $macro,
             'init_method' => $init_method,
             'associated_content' => $associated_content_saved,
             'errors' => $errors,
-            'none_option' => ['0' => $macro->letter_type->name === 'Internal Referral' ? 'Internal Referral' : 'None'],
-            'recipients_data' => \CHtml::listData(LetterRecipient::model()->findAll(array('order' => 'display_order asc')), 'id', 'name'),
-            'label_options' => [
-                $letter_type_gp_id => $macro->letter_type->name === 'Internal Referral' ? 'display:none' : '',
-                $letter_type_patient_id => $macro->letter_type->name === 'Internal Referral' ? 'display:none' : ''
-            ],
         ));
     }
 
     public function actionDeleteLetterMacros()
     {
+        if(!isset($_POST['id'])) {
+            return null;
+        }
+        //Make all the macro ids null that is equal to the macro id
+        // that is being deleted in the document instance data table
+        DocumentInstanceData::model()->updateAll(['macro_id' => null], 'macro_id IN (' . implode($_POST['id']) . ')');
+
         $criteria = new CDbCriteria();
-
         $criteria->addInCondition('id', @$_POST['id']);
-
         if (LetterMacro::model()->deleteAll($criteria)) {
             echo '1';
         } else {
